requirements.txt
allure-pytest==2.8.29
allure-python-commons==2.8.29
apipkg==1.5
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
chevron==0.14.0
cycler==0.10.0
execnet==1.7.1
idna==2.10
iniconfig==1.1.1
jsonschema==3.2.0
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
packaging==20.8
Pillow==8.0.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pyrsistent==0.17.3
pystache==0.5.4
pytest==6.2.1
pytest-assume==2.4.2
pytest-base-url==1.4.2
pytest-forked==1.3.0
pytest-xdist==2.2.0
python-dateutil==2.8.1
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.3
